requirements.txt
openai
tqdm
pandas

python-docx
PyPDF2

pytest
